src/main/java/com/sencerseven/basittarifler/converter/IngredientToIngredientCommandConverter.java
package com.sencerseven.basittarifler.converter;

import com.sencerseven.basittarifler.command.IngredientCommand;
import com.sencerseven.basittarifler.domain.Ingredient;
import org.springframework.core.convert.converter.Converter;
import org.springframework.stereotype.Component;

@Component
public class IngredientToIngredientCommandConverter implements Converter<Ingredient,IngredientCommand> {


    @Override
    public IngredientCommand convert(Ingredient ingredient) {
        if(ingredient == null)
            return  null;
        IngredientCommand ingredientCommand = new IngredientCommand();

        ingredientCommand.setId(ingredient.getId());
        ingredientCommand.setDescription(ingredient.getDescription());

        return ingredientCommand;


    }
}
